Add CAS-based one callback at a time and promise linking implementations #32
The promise linking does not have compression yet and therefore is slower
and leads to a stack overflow in the recursion of too many calls (see TODOs).

Benchmark 5 with the call of perf4Prim etc. tests recursive calls which lead
to a big chain of promises.
With this benchmark we can compare the effects of promise linking on the performance.

Simplify the specializations of CallbackEntry and support ParentCallbackEntry
which is called ManyCallbacks in the Scala standard library.